- Added BookingTableServiceImpl with complete CRUD methods: add, getAll, getById, delete, update

+++ src/main/java/com/shreya/hibernate/repository/BookingTableRepository.java
@@ -4,6 +4,6 @@ import com.shreya.hibernate.domain.BookingTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BookingTableRepository extends JpaRepository<BookingTable, Integer> {
+public interface BookingTableRepository extends JpaRepository<BookingTable, Long> {
 
 }
